+++ addon/data-import.js
@@ -1066,12 +1066,12 @@ class StatusBox extends React.Component {
   initButton(sfHost, true);
   sfConn.getSession(sfHost).then(() => {
 
-    let root = document.getElementById("root");
+    let root = ReactDOM.createRoot(document.getElementById("root"));
     let model = new Model(sfHost);
     model.reactCallback = cb => {
-      ReactDOM.render(h(App, {model}), root, cb);
+      root.render(h(App, {model}));
     };
-    ReactDOM.render(h(App, {model}), root);
+    root.render(h(App, {model}));
 
     if (parent && parent.isUnitTest) { // for unit tests
       parent.insextTestLoaded({model});