+++ addon/explore-api.js
@@ -324,7 +324,7 @@ class App extends React.Component {
   initButton(sfHost, true);
   sfConn.getSession(sfHost).then(() => {
 
-    let root = document.getElementById("root");
+    let root = ReactDOM.createRoot(document.getElementById("root"));
     let model = new Model(sfHost, args);
     window.sfConn = sfConn;
     window.display = apiPromise => {
@@ -334,9 +334,9 @@ class App extends React.Component {
       model.performRequest(Promise.resolve(apiPromise));
     };
     model.reactCallback = cb => {
-      ReactDOM.render(h(App, {model}), root, cb);
+      root.render(h(App, {model}));
     };
-    ReactDOM.render(h(App, {model}), root);
+    root.render(h(App, {model}));
 
   });